lib: make use of __VA_OPT__ with C23 and C++20 (and later) compilers
We use a `##__VA_ARGS__` solution to work with zero arguments with
variadic macros. But this is not a standard C/C++ feature and is
a language extension, which breaks when using stricter
standards-compliant compilation modes, like --std=c23 or --std=c++2ai.

Luckily, the standard way now is to use __VA_OPT__(,) to achieve the same.

Unluckily, this requires a bunch of duplication in 4 different places.

Previously, the following would result in hard to understand compilation
error because it would count empty __VA_ARGS__ as having one argument:

make CXXFLAGS=--std=c++2a V=1 clean build-cxx_arg_nums

With this change it now works both with --std=c++2a/--std=c23 and without.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>