tests: avoid sending signals in run_test.sh
Change run_test.sh and .bt script such that bpftrace terminates at the
end of target test program. Compare the test output after bpftrace has
terminated.

Using kill -INT to stop bpftrace may interfere with CI environment,
because this signal may be interpreted as a job failure.

I tried bpftrace -c [1] to work around this, however it turned out to
not work for these tests when there are tracepoints refering to
dynamic libraries. This is because the .so is not loaded yet when
bpftrace attempts to attach to the child process.

[1] https://docs.bpftrace.org/latest#-c-command

Signed-off-by: Ihor Solodrai <ihor.solodrai@pm.me>