tests: skip a faulty test case when compiling with clang
The "struct_by_val_reg" case of arg_types tests fails with BPF
verifier error when compiled with clang.

It appears that bpftrace generates a faulty program in such case,
possibly due to the difference between how clang and gcc encode USDT
probe arguments.

Since this is the only problematic test case for LLVM, make it
conditional on __clang__ for now.

Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>